InteractionMap class
used for MiscEntity and NPCEntity
